Works. Added documentation.
