+++ Documentation/python/add_ons/add_on.md
@@ -39,6 +39,7 @@ c.communicate({"$type": "terminate"})
 - [Keyboard](keyboard.md)
 - [ObjectManager](object_manager.md)
 - [OccupancyMap](occupancy_map.md)
+- [PhysicsAudioRecorder](physics_audio_recorder.md)
 - [Robot](robot.md)
 - [StepPhysics](step_physics.md)
 - [ThirdPersonCamera](third_person_camera.md)
