+++ Documentation/python/add_ons/physics_audio_recorder.md
@@ -1,0 +1,93 @@
+# PhysicsAudioRecorder
+
+`from tdw.add_ons.physics_audio_recorder import PhysicsAudioRecorder`
+
+Record audio from a TDW build.
+
+***
+
+## Fields
+
+- `commands` These commands will be appended to the commands of the next `communicate()` call.
+
+- `initialized` If True, this module has been initialized.
+
+- `max_frames` # If greater than 0, stop recording after this many frames even if objects are still moving or making sound.
+
+- `output_path` The path to the next audio file.
+
+- `recording` If True, there is an ongoing recording.
+
+***
+
+## Functions
+
+#### \_\_init\_\_
+
+**`PhysicsAudioRecorder()`**
+
+**`PhysicsAudioRecorder(max_frames=-1)`**
+
+| Parameter | Type | Default | Description |
+| --- | --- | --- | --- |
+| max_frames |  int  | -1 | If greater than 0, stop recording after this many frames even if objects are still moving or making sound. |
+
+#### get_initialization_commands
+
+**`self.get_initialization_commands()`**
+
+This function gets called exactly once per add-on. To re-initialize, set `self.initialized = False`.
+
+_Returns:_  A list of commands that will initialize this add-on.
+
+#### on_send
+
+**`self.on_send(output_path)`**
+
+Start recording.
+
+| Parameter | Type | Default | Description |
+| --- | --- | --- | --- |
+| output_path |  |  | The path to the output .wav file. |
+
+#### start
+
+**`self.start(output_path)`**
+
+Start recording.
+
+| Parameter | Type | Default | Description |
+| --- | --- | --- | --- |
+| output_path |  Union[str, Path] |  | The path to the output .wav file. |
+
+#### stop
+
+**`self.stop()`**
+
+Stop an ongoing recording. Use ffmpeg to remove initial silence.
+
+#### on_send
+
+**`self.on_send(resp)`**
+
+This is called after commands are sent to the build and a response is received.
+
+Use this function to send commands to the build on the next frame, given the `resp` response.
+Any commands in the `self.commands` list will be sent on the next frame.
+
+| Parameter | Type | Default | Description |
+| --- | --- | --- | --- |
+| resp |  List[bytes] |  | The response from the build. |
+
+#### before_send
+
+**`self.before_send(commands)`**
+
+This is called before sending commands to the build. By default, this function doesn't do anything.
+
+| Parameter | Type | Default | Description |
+| --- | --- | --- | --- |
+| commands |  List[dict] |  | The commands that are about to be sent to the build. |
+
+
+
