+++ Documentation/python/model_tests/rotate_object_test.md
@@ -12,7 +12,9 @@ These tests add an object and then rotate it.
 | --- | --- | --- |
 | `OBJECT_ID` | int | The ID of the object. |
 | `DELTA_THETA` | int | Rotate by this many degrees per frame. |
-| `PINK ` |  | The Unity pink color. |
+| `PINK` | tuple | The Unity pink color. |
+| `LOOK_AT` | Dict[str, float] | Look at this position. |
+| `AVATAR_POSITION` | Dict[str, float] | The position of the avatar. |
 
 ***
